+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## TODO
 
+* Add react-router for navigation 
+* Add responsive views
+* Fix markdown loading (each .md file loads 3x per button click)
 * Create/restructure new content for all empty .md pages
 * Deprecate Expressions section and integrate content with Logic section. Separate documentation into logical chunks according to sub-topic.
 * Deprecate Response Types section and integrate content with Basic Operations section.
@@ -13,11 +16,18 @@
 * MUI Styles updates for various components
 
 ---
+## v.1.5 (2021-01-17)
+
+### Added
+
+* Responsive drawer
+
+---
 
 ## 1.0 (2021-01-12)
 
 ### Added
- 
+
  * Markdown files added (complete and incomplete) to indicate general skeleton of documentation
  * Additional Typescript migration